+++ tests/testthat/test-Plot2DGraph.R
@@ -5,6 +5,7 @@ for (assay_version in c("v3", "v5")) {
   seur_obj <- ReadMPX_Seurat(pxl_file, overwrite = TRUE)
   seur_obj <- LoadCellGraphs(seur_obj, cells = colnames(seur_obj)[1:2])
   seur_obj <- ComputeLayout(seur_obj, layout_method = "pmds")
+  seur_obj <- ComputeLayout(seur_obj, layout_method = "cpmds")
 
   test_that("Plot2DGraph works as expected", {
     expect_no_error({
@@ -17,6 +18,10 @@ for (assay_version in c("v3", "v5")) {
     expect_equal(layout_plot[[1]]$labels$title, "RCVCMP0000217")
     expect_equal(layout_plot[[1]]$labels$colour, "CD3E\n(log-scaled)")
     expect_equal(dim(layout_plot[[1]]$data), c(2470, 8))
+
+    expect_no_error({
+      layout_plot <- Plot2DGraph(seur_obj, cells = colnames(seur_obj)[1], layout_method = "cpmds_3d")
+    })
 
     # Test with showBnodes active
     expect_no_error({
@@ -62,9 +67,13 @@ for (assay_version in c("v3", "v5")) {
     })
     expect_s3_class(layout_plot, "patchwork")
     expect_equal(layout_plot[[1]]$labels$title, NULL)
+
+    expect_no_error({
+      layout_plot <- Plot2DGraphM(seur_obj, cells = colnames(seur_obj)[1], layout_method = "cpmds_3d", markers = c("HLA-DR"))
+    })
   })
 
-  test_that("Plot2DGraph fails with invalid input", {
+  test_that("Plot2DGraphM fails with invalid input", {
     expect_error({
       layout_plot <- Plot2DGraphM(seur_obj, cells = colnames(seur_obj)[1], layout_method = "invalid")
     })